refact, docs
Fixed camera transformation; Improved readability

The camera transformation has changed from the RobustScaler to the StandardScaler for a better distribution of the values. For this reason new training and test datasets have been created with the new adapted transformation.
The readability of the notebook has been improved for presentation purposes. New histograms and plots explaining the reasons of some processing techinques used.